tests/wpt/web-platform-tests/css/css-values/minmax-percentage-serialize.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#percentages">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-serialize">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/parsing-testcommon.js"></script>
<script>
function test_valid_percentage(value, expected) {
  test_valid_value('margin-left', value, expected);
}

test_valid_percentage('min(1%)', 'min(1%)');
test_valid_percentage('max(1%)', 'max(1%)');

test_valid_percentage('min(1%, 2%, 3%)', 'min(1%, 2%, 3%)');
test_valid_percentage('min(3%, 2%, 1%)', 'min(3%, 2%, 1%)');
test_valid_percentage('max(1%, 2%, 3%)', 'max(1%, 2%, 3%)');
test_valid_percentage('max(3%, 2%, 1%)', 'max(3%, 2%, 1%)');

test_valid_percentage('calc(min(1%) + min(2%))', 'calc(min(1%) + min(2%))');
test_valid_percentage('calc(max(1%) + max(2%))', 'calc(max(1%) + max(2%))');
test_valid_percentage('calc(1% + min(1%))', 'calc(1% + min(1%))');
test_valid_percentage('calc(min(1%) + 1%)', 'calc(1% + min(1%))');
test_valid_percentage('calc(1% + max(1%))', 'calc(1% + max(1%))');
test_valid_percentage('calc(max(1%) + 1%)', 'calc(1% + max(1%))');

</script>
